requirements.txt
chroma-hnswlib==0.7.3
chromadb==0.5.3
langchain==0.2.6
langchain_community==0.2.6
langchain_openai==0.1.12
langchain_text_splitters==0.2.2
python-dotenv==1.0.1
PyYAML==6.0.1
streamlit==1.36.0
streamlit_authenticator==0.3.2
streamlit_pdf_viewer==0.0.14
sentence-transformers
openai
pymupdf==1.24.7
pymupdfb==1.24.6
pypdf